# docs/swagger.yaml
definitions:
...
  dto.LoginUserDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.RegisterUserDTO:
    properties:
      country:
        type: string
      email:
        type: string
      fullname:
        maxLength: 50
        minLength: 2
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      password:
        minLength: 6
        type: string
      phone_number:
        maxLength: 15
        minLength: 7
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - country
    - email
    - fullname
    - gender
    - password
    - phone_number
    - username
    type: object
  dto.ResetPasswordDTO:
    properties:
      confirm_password:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - confirm_password
    - password
    - token
    type: object
  dto.ResetPasswordLinkDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.UpdatePasswordDTO:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  dto.UpdateUserBalanceDTO:
    properties:
      invested_amount:
        type: number
      profit_balance:
        type: number
      referal_bonus:
        type: number
      wallet_balance:
        type: number
    required:
    - invested_amount
    - profit_balance
    - referal_bonus
    - wallet_balance
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is my Go Inv API.
  title: My Inv API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      description: Fetch all user data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch all users
      tags:
      - Admin Dashboard
  /admin/users/{id}:
    get:
      description: Fetch user data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Fetch user
      tags:
      - Admin Dashboard
  /admin/users/{id}/delete:
    delete:
      consumes:
      - application/json
      description: delete user account as an admin
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - Admin Action
  /admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      description: update user account status as an admin
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user account status
      tags:
      - Admin Action
  /admin/users/{id}/update:
    put:
      consumes:
      - application/json
      description: update user account as an admin
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserBalanceDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user account
      tags:
      - Admin Action
...
  /create:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, fullname, phone, and password
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Register a new user
      tags:
      - Authentication
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the user's email
      parameters:
      - description: Password reset request data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordLinkDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Request password reset link
      tags:
      - Authentication
  /login:
    post:
      consumes:
      - application/json
      description: Login user account using email and password
      parameters:
      - description: login user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Login User
      tags:
      - Authentication
  /refresh:
    post:
      consumes:
      - application/json
      description: verify refresh token and send back a new access token
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Refresh token
      tags:
      - Authentication
  /reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password from token
      parameters:
      - description: Password reset request data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Reset user password
      tags:
      - Authentication
...
  /user/profiles:
    get:
      description: Returns user profile page data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User profile page data
      tags:
      - User Dashboard
...
  /user/update-password:
    patch:
      consumes:
      - application/json
      description: update user password from profile settings
      parameters:
      - description: Password Update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user password
      tags:
      - User Action
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"